requirements.txt
web3==7.14.1
google-generativeai==0.8.6
anthropic==0.84.0
requests==2.32.5
google-cloud-secret-manager==2.26.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
fastmcp==0.4.1
EOF"